tests/wpt/tests/html/canvas/element/text/2d.text.drawing.style.reset.fontKerning.none.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.drawing.style.reset.fontKerning.none</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.reset.fontKerning.none</h1>
<p class="desc">crbug/338965374, fontKerning still works after setting font for a second time.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("crbug/338965374, fontKerning still works after setting font for a second time.");
_addTest(function(canvas, ctx) {

  ctx.font = '100px serif';
  ctx.fontKerning = "none";
  const width1 = ctx.measureText("AW").width;
  ctx.font = '100px serif';
  _assertSame(ctx.fontKerning, "none", "ctx.fontKerning", "\"none\"");
  const width2 = ctx.measureText("AW").width;
  _assertSame(width1, width2, "width1", "width2");

});
</script>
